sunbeam_ai_bot/requirements.txt
streamlit
selenium
chromadb
langchain
python-dotenv
webdriver-manager
